+++ Assets/Scripts/Unity/TrialParameters.cs
@@ -4,14 +4,14 @@ using UnityEngine;
 [System.Serializable]
 public class TrialParameters
 {
-    public int roughness_left;
+    public float roughness_left;
     public float amplitude_left;
     public float frequency_left;
-    public int roughness_right;
+    public float roughness_right;
     public float amplitude_right;
     public float frequency_right;
     public TrialParameters(){}
-    public TrialParameters(int rl, float al, float pl, int rr, float ar, float pr){
+    public TrialParameters(float rl, float al, float pl, float rr, float ar, float pr){
             roughness_left = rl;
             amplitude_left = al;
             frequency_left = pl;
